Update React Examples based on changes to React
**what is the change?:**
We are deprecating the 'componentWillMount' method in React, and
providing new patterns and helpers for doing subscription and other
common patterns which used to rely on 'componentWillMount'.

**why make this change?:**
If folks copy the current implementation in the examples, they will get
warnings, and with 'async mode' activated they will encounter subtle
bugs. See <TODO>

**test plan:**
WIP - need to run the browser test, but for now did manual testing and
things still work fine.

**issue:**
WIP